Create Year.pkg
Happy New Year!
As a gift, here's a new app I made: Year.
It's a simple journaling app, but every day, you can access ONLY that text.

Plan to add a feature to review previous posts, like at the end of the year. Also want to make a widget to remind you daily.