Oracular Oriole Wallpapers
Ubuntu Oracular Oriole wallpapers
Uploads of several wallpapers I created. The Oriole Mascot (Dark, Light) will be present in the official release. The 6k Dynamic is an expanded version of that wallpaper, with a 6016x6016 resolution .

The Ubuntu 20 wallpapers, celebrating 20 years of Ubuntu, is only on my website (finished it like 4 hours later than the deadline 'cause of time zones, lmao)